+++ eval_reasoning.py
@@ -19,24 +19,13 @@ This logic is largely copied from the Hendrycks' MATH release (math_equivalence)
 
 import argparse
 import json
-import multiprocessing
 import os
-import queue
 import random
 import re
 from collections import Counter
-from math import isclose
-from typing import Union
 
-import evaluate
-import regex
 import torch
 from datasets import load_dataset
-from latex2sympy2 import latex2sympy
-from sympy import N
-from sympy import simplify
-from sympy.parsing.latex import parse_latex
-from sympy.parsing.sympy_parser import parse_expr
 from tqdm import tqdm
 from tqdm import trange
 from transformers import AutoModelForCausalLM
@@ -47,340 +36,15 @@ from nncf.quantization.advanced_parameters import KVCacheCompressionParameters
 from nncf.quantization.advanced_parameters import KVCacheRefinedSelection
 from nncf.quantization.algorithms.kv_cache_management.torch_backend import KVCacheCompressor
 from reasoning_parser import extract_answer
+from reasoning_parser import parallel_math_equal
 from reasoning_parser import strip_string
 
 # disable tokenizer parallelism warnings
 os.environ["TOKENIZERS_PARALLELISM"] = "false"
 OUTPUT_LENGTHS = []
 
-exact_match = evaluate.load("exact_match")
 
-
-def choice_answer_clean(pred: str):
-    pred = pred.strip("\n").rstrip(".").rstrip("/").strip(" ").lstrip(":")
-    # Clean the answer based on the dataset
-    tmp = re.findall(r"\b(A|B|C|D|E)\b", pred.upper())
-    if tmp:
-        pred = tmp
-    else:
-        pred = [pred.strip().strip(".")]
-    pred = pred[-1]
-    # Remove the period at the end, again!
-    pred = pred.rstrip(".").rstrip("/")
-    return pred
-
-
-def parse_digits(num):
-    num = regex.sub(",", "", str(num))
-    try:
-        return float(num)
-    except ValueError:
-        if num.endswith("%"):
-            num = num[:-1]
-            if num.endswith("\\"):
-                num = num[:-1]
-            try:
-                return float(num) / 100
-            except ValueError:
-                pass
-    return None
-
-
-def is_digit(num):
-    # paired with parse_digits
-    return parse_digits(num) is not None
-
-
-def str_to_pmatrix(input_str):
-    input_str = input_str.strip()
-    matrix_str = re.findall(r"\{.*,.*\}", input_str)
-    pmatrix_list = []
-
-    for m in matrix_str:
-        m = m.strip("{}")
-        pmatrix = r"\begin{pmatrix}" + m.replace(",", "\\") + r"\end{pmatrix}"
-        pmatrix_list.append(pmatrix)
-
-    return ", ".join(pmatrix_list)
-
-
-def math_equal(
-    prediction: Union[bool, float, str],
-    reference: Union[float, str],
-    include_percentage: bool = True,
-    is_close: bool = True,
-    timeout: bool = False,
-) -> bool:
-    """
-    Exact match of math if and only if:
-    1. numerical equal: both can convert to float and are equal
-    2. symbolic equal: both can convert to sympy expression and are equal
-    """
-    # print("Judge:", prediction, reference)
-    if prediction is None or reference is None:
-        return False
-    if str(prediction.strip().lower()) == str(reference.strip().lower()):
-        return True
-    if reference in ["A", "B", "C", "D", "E"] and choice_answer_clean(prediction) == reference:
-        return True
-
-    try:  # 1. numerical equal
-        if is_digit(prediction) and is_digit(reference):
-            prediction = parse_digits(prediction)
-            reference = parse_digits(reference)
-            # number questions
-            if include_percentage:
-                gt_result = [reference / 100, reference, reference * 100]
-            else:
-                gt_result = [reference]
-            for item in gt_result:
-                try:
-                    if is_close:
-                        if numeric_equal(prediction, item):
-                            return True
-                    else:
-                        if item == prediction:
-                            return True
-                except Exception:
-                    continue
-            return False
-    except Exception:
-        pass
-
-    if not prediction and prediction not in [0, False]:
-        return False
-
-    # 2. symbolic equal
-    reference = str(reference).strip()
-    prediction = str(prediction).strip()
-
-    # pmatrix (amps)
-    if "pmatrix" in prediction and "pmatrix" not in reference:
-        reference = str_to_pmatrix(reference)
-
-    # deal with [], (), {}
-    pred_str, ref_str = prediction, reference
-    if (prediction.startswith("[") and prediction.endswith("]") and not reference.startswith("(")) or (
-        prediction.startswith("(") and prediction.endswith(")") and not reference.startswith("[")
-    ):
-        pred_str = pred_str.strip("[]()")
-        ref_str = ref_str.strip("[]()")
-    for s in ["{", "}", "(", ")"]:
-        ref_str = ref_str.replace(s, "")
-        pred_str = pred_str.replace(s, "")
-    if pred_str.lower() == ref_str.lower():
-        return True
-
-    # [a, b] vs. [c, d], return a==c and b==d
-    if (
-        regex.match(r"(\(|\[).+(\)|\])", prediction) is not None
-        and regex.match(r"(\(|\[).+(\)|\])", reference) is not None
-    ):
-        pred_parts = prediction[1:-1].split(",")
-        ref_parts = reference[1:-1].split(",")
-        if len(pred_parts) == len(ref_parts) and all(
-            [math_equal(pred_parts[i], ref_parts[i], include_percentage, is_close) for i in range(len(pred_parts))]
-        ):
-            return True
-    if (
-        (prediction.startswith("\\begin{pmatrix}") or prediction.startswith("\\begin{bmatrix}"))
-        and (prediction.endswith("\\end{pmatrix}") or prediction.endswith("\\end{bmatrix}"))
-        and (reference.startswith("\\begin{pmatrix}") or reference.startswith("\\begin{bmatrix}"))
-        and (reference.endswith("\\end{pmatrix}") or reference.endswith("\\end{bmatrix}"))
-    ):
-        pred_lines = [
-            line.strip()
-            for line in prediction[len("\\begin{pmatrix}") : -len("\\end{pmatrix}")].split("\\\\")
-            if line.strip()
-        ]
-        ref_lines = [
-            line.strip()
-            for line in reference[len("\\begin{pmatrix}") : -len("\\end{pmatrix}")].split("\\\\")
-            if line.strip()
-        ]
-        matched = True
-        if len(pred_lines) == len(ref_lines):
-            for pred_line, ref_line in zip(pred_lines, ref_lines):
-                pred_parts = pred_line.split("&")
-                ref_parts = ref_line.split("&")
-                if len(pred_parts) == len(ref_parts):
-                    if not all(
-                        [
-                            math_equal(
-                                pred_parts[i],
-                                ref_parts[i],
-                                include_percentage,
-                                is_close,
-                            )
-                            for i in range(len(pred_parts))
-                        ]
-                    ):
-                        matched = False
-                        break
-                else:
-                    matched = False
-                if not matched:
-                    break
-        else:
-            matched = False
-        if matched:
-            return True
-
-    if prediction.count("=") == 1 and reference.count("=") == 1:
-        pred = prediction.split("=")
-        pred = f"{pred[0].strip()} - ({pred[1].strip()})"
-        ref = reference.split("=")
-        ref = f"{ref[0].strip()} - ({ref[1].strip()})"
-        if symbolic_equal(pred, ref) or symbolic_equal(f"-({pred})", ref):
-            return True
-    elif prediction.count("=") == 1 and len(prediction.split("=")[0].strip()) <= 2 and "=" not in reference:
-        if math_equal(prediction.split("=")[1], reference, include_percentage, is_close):
-            return True
-    elif reference.count("=") == 1 and len(reference.split("=")[0].strip()) <= 2 and "=" not in prediction:
-        if math_equal(prediction, reference.split("=")[1], include_percentage, is_close):
-            return True
-
-    # symbolic equal with sympy
-    if timeout:
-        if call_with_timeout(symbolic_equal_process, prediction, reference):
-            return True
-    else:
-        if symbolic_equal(prediction, reference):
-            return True
-
-    return False
-
-
-def math_equal_process(param):
-    return math_equal(param[-2], param[-1])
-
-
-def numeric_equal(prediction: float, reference: float):
-    # Note that relative tolerance has significant impact
-    # on the result of the synthesized GSM-Hard dataset
-    # if reference.is_integer():
-    #     return isclose(reference, round(prediction), abs_tol=1e-4)
-    # else:
-    # prediction = round(prediction, len(str(reference).split(".")[-1]))
-    return isclose(reference, prediction, rel_tol=1e-4)
-
-
-def symbolic_equal(a, b):
-    def _parse(s):
-        for f in [parse_latex, parse_expr, latex2sympy]:
-            try:
-                return f(s.replace("\\\\", "\\"))
-            except Exception:
-                try:
-                    return f(s)
-                except Exception:
-                    pass
-        return s
-
-    a = _parse(a)
-    b = _parse(b)
-
-    # direct equal
-    try:
-        if str(a) == str(b) or a == b:
-            return True
-    except Exception:
-        pass
-
-    # simplify equal
-    try:
-        if a.equals(b) or simplify(a - b) == 0:
-            return True
-    except Exception:
-        pass
-
-    # equation equal
-    try:
-        if (abs(a.lhs - a.rhs)).equals(abs(b.lhs - b.rhs)):
-            return True
-    except Exception:
-        pass
-
-    try:
-        if numeric_equal(float(N(a)), float(N(b))):
-            return True
-    except Exception:
-        pass
-
-    # matrix
-    try:
-        # if a and b are matrix
-        if a.shape == b.shape:
-            _a = a.applyfunc(lambda x: round(x, 3))
-            _b = b.applyfunc(lambda x: round(x, 3))
-            if _a.equals(_b):
-                return True
-    except Exception:
-        pass
-
-    return False
-
-
-def symbolic_equal_process(a, b, output_queue):
-    result = symbolic_equal(a, b)
-    output_queue.put(result)
-
-
-def call_with_timeout(func, *args, timeout=1, **kwargs):
-    output_queue = multiprocessing.Queue()
-    process_args = args + (output_queue,)
-    process = multiprocessing.Process(target=func, args=process_args, kwargs=kwargs)
-    process.start()
-    process.join(timeout)
-
-    if process.is_alive():
-        process.terminate()
-        process.join()
-        return False
-
-    return output_queue.get()
-
-
-def math_equal_with_timeout(pred, gt_ans, timeout):
-    def target(result_queue):
-        try:
-            result_queue.put(math_equal(pred, gt_ans))
-        except Exception as e:
-            result_queue.put(e)
-
-    result_queue = multiprocessing.Queue()
-    process = multiprocessing.Process(target=target, args=(result_queue,))
-    process.start()
-
-    process.join(timeout)
-
-    if process.is_alive():
-        print(f"Timeout occurred for prediction: {pred}")
-        process.terminate()
-        process.join()
-        return False
-
-    try:
-        result = result_queue.get_nowait()
-    except queue.Empty:
-        print("Result queue timed out")
-        return False
-
-    if isinstance(result, Exception):
-        print(f"Error occurred: {result}")
-        return False
-
-    return result
-
-
-def parallel_math_equal(all_pred, gt_ans, timeout=20):
-    results = []
-    for pred in all_pred:
-        results.append(math_equal_with_timeout(pred, gt_ans, timeout))
-    return results
-
-
-def eval_main(res_path, save=False, k=None, output_dir=None):
+def run_evaluation(res_path, save=False, k=None, output_dir=None):
     with open(res_path) as f:
         lines = f.readlines()
         data = [json.loads(line) for line in lines]
@@ -472,16 +136,6 @@ def extract_box(pred_str):
         a = ans.split("$")[0].strip()
 
     return a
-
-
-def extract_last_number(pred_str):
-    o = re.sub(r"(\d),(\d)", r"\1\2", pred_str)
-    numbers = re.findall(r"[-+]?\d*\.\d+|\d+", o)
-    if numbers:
-        ans = numbers[-1]
-    else:
-        ans = None
-    return ans
 
 
 def main(args):
@@ -686,4 +340,4 @@ if __name__ == "__main__":
 
     print(args.save_dir)
     main(args)
-    eval_main(os.path.join(args.save_dir, "predictions.jsonl"), save=True, k=None, output_dir=args.save_dir)
+    run_evaluation(os.path.join(args.save_dir, "predictions.jsonl"), save=True, k=None, output_dir=args.save_dir)
